LayoutTests/http/tests/appcache/interrupted-update-expected.txt

CONSOLE MESSAGE: line 1: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: line 37: Reloading...
CONSOLE MESSAGE: line 1: ApplicationCache is deprecated. Please use ServiceWorkers instead.
Test that an application cache update works again after interruption.

Reloaded
CHECKING
DOWNLOADING
PROGRESS 0/19
PROGRESS 1/19
PROGRESS 2/19
PROGRESS 3/19
PROGRESS 4/19
PROGRESS 5/19
PROGRESS 6/19
PROGRESS 7/19
PROGRESS 8/19
PROGRESS 9/19
PROGRESS 10/19
PROGRESS 11/19
PROGRESS 12/19
PROGRESS 13/19
PROGRESS 14/19
PROGRESS 15/19
PROGRESS 16/19
PROGRESS 17/19
PROGRESS 18/19
PROGRESS 19/19
CACHED

PASS
